v0.3.0 - Audio Settings
Added Audio Settings - the user can adjust the volume level or mute the sound.
Audio settings are saved in the loaclStorage, so the they will be set automatically to the user settings when revisiting the page.

The time buttons (game duration) are now disabled after the first game (the time buttons in the final screen are still enabled).