v0.4.0 - Scoreboard & Personal Best Score
Added a Scoreboard and Personal Best Score.
It's tracking the most important stats (for each mode - game duration) and keeps them saved in the localStorage.
It can calculate average values (for each mode). The user can clear the scoreboard.

There is also the Personal Best Score feature. It shows the best score from the scoreboard (not from the first game and not if it's the same as the previous highest score).
It gets cleared when the user clears the scoreboard.