v0.5.1 - Music Player Controls
Added controls to the music player:
- play / pause
- prev / next track
- mute
- volume adjustment

The music will automatically play after the preparation timer, unless the user first pauses it.
The music can be played without starting the game (by clicking "unpause" button).

Volume and mute are saved in the localStorage.
The current track is also saved in the localStorage, so the last track will be played when starting a new game session (after entering the page).