v0.7.1 - Game Theme
Added changing game theme based on the level achieved by the user. Each time the game is finished, the game theme will change (it won't stay the same based on the highest score to encourage players to perform better).

If the user ends the game manually (finish button), the game theme will be set to default to discourage players from restarting the game all the time - the focus of the game is not only to train precision, but also consistency.