v0.7.2 - Logo, Favicon, Typefaces
Designed a new 3d logo, favicon and chose typefaces.

The logo is 3d - there are shadows that are calculated by subtracting values from the main color (CSS variables).
Because of that, the shadows respond to the main color change when the user achieves a certain level (more in the v0.7.1 update).