v0.8.0 - Background Blobs, Basic CSS, Color Calc
Added Background Blobs - randomly generated, colorful blobs (color depends on the current theme) that move across the screen. They are "breathing" and rotating infinitely - there are 3 groups, each one has a slightly different animation duration and delay, so blobs seem more unique.
They are additionally blurred behing the main wrapper elements.

The blobs react to the screen - they won't go out of view when resizing the page, but they will be dragged with the screen edge to still be in the view.
The number of blobs dynamically changes based on the screen width.

Added stylesheets with some bacis styles for the page, like css variables (to standardize the design), text, wrapper, etc. - basically, they are only placeholders that will change.

The background color is a gradient that reacts to the theme color change.

Improved the Color Calculator - now, it can take more properties, like to change different css variable colors, adding transparency, and adding or subtracting values.

+ Many more minor changes and adjustments.