v0.8.1 - No Score (Inactive Player)
If the player has a score of 0, the theme will be reset to default and there won't be any level info, such as fun facts etc. The player needs at least 1 point for the game to count.

If the score is 0, the game results won't be saved to the scoreboard - the game will be treated as if it never happened.