v0.9.0 - Design, Background Stars, Performance
Added frosted glass look to the elements. It's based on the current theme, the colors change with the achieved level.

Improved the Game Field design (including the start button and the preparation timer). The player has more info about the current action (hover effects).

Changed the moving objects (Background Blobs and the Game Core) from "translate" to "translate3d" for better performance and smoother animations.
Improved positioning, especially for the Game Core, now it's more accurate and has the full range of pixels inside the field.

Updated the background stars - now, some of them will be rotated 45 deg and will be the main color, so the background feels more unique and random.

+ More minor changes.