v0.10.0 - Game Termination, Core, Audio Players
Added a feature that terminates the game when the user changes the window size.
It's designed to prevent unexpected behavior, such as changes in layout and cheating.

Added messages displayed when the game was finished early / with no score:
- game terminated by the user ("finish" button clicked)
- game terminated when resizing the browser window
- no score
In each of these examples, the results won't be saved to the scoreboard.

Added dynamic core size adjustment based on the game field size (container queries).
Changed the core animation when changing its position.
Now, it doesn't have a transition animation to a new position. Instead, it appears immediately and scales itself from nothing to its normal size (grows).
Because of those changes, the visuals are now more pleasant to look at and the transition is less laggy.
It also fixed a bug where the core could be seen transitioning from the top-left corner to its position immediately after the game start.

Added a new indicator design to the current game duration button (accent color).
Added an indicator to the game duration buttons to show when they are disabled.

Added styling to the audio player elements, including marquee for too long music titles.
Designed new icons for the Audio Player buttons.
The dynamic mute icon changes it's appearance based on the volume set by the user.

Changed the cursor when it's in the Game Field during the game to a custom crosshair that changes colors based on the current theme (SVG).

Changed the scrollbar color.