v0.11.0 - Main Screen, Media Query, Core Design
Changed the core design to fit the game's aesthetics and to utilize the current theme.

Added media queries to the main screen.

Fixed the bug where the game duration buttons' gradients did not transition correctly (gradients do not transition).

Implemented Game Field aspect ratio adjustment based on the screen size.
Now, it will always fit the screen and optimally utilizes the device's aspect ratio.

Designed icons for the score and time.
They can react to the game events, such as the time running out and when the player misclicks.

Fixed a bug where the core was registered as clicked when the user clicked outside the circle, behaving as if the core were square.

Fixed minor bugs and implemented minor design and script adjustments.