v1.0.1 - Version, Design Fixes
Changed the version script to source it from the "package.json" and display it in the footer.

Implemented some design fixes in the "Stats" section.
Changed the paddings and positioning of certain elements to improve readability, consistency, and to make the design pixel-perfect.

Enhanced the fluidity of the marquee animation in the music player.